Version: 4.3
Solved a critical problem with the bottom page section for the front page.
Split the advanced settings panel into several panels because it was becoming too large
and difficult to navigate.
Added an option to include a small version of the logo in the menu.
Updated the default colors of the Call to Action to improve accessibility.